Add .gitattributes
The line endings of text files in this repo are not consistent and some
files even have mixed line endings, presumably from being edited with an
editor that neither respects a file's existing EOL format nor converts
existing line endings on save.

This commit adds a .gitattributes file to mark most text files as text
in Git, which will cause Git to normalize the line endings in these
files to LF internally and to the format specified by the core.eol
setting for files in the working tree.

This commit only adds .gitattributes and does not change line endings.
The following commit normalizes all line endings for files given the
`text` attribute by this commit.